+++ tests/cofi_tools/test_all_tools.py
@@ -1,6 +1,6 @@
-# ############################### TEST ALL SOLVERS ###############################
-# 1. The solver is added to solvers/__init__.py 
-#        __all__, and solvers_table
+# ############################### TEST ALL TOOLS ###############################
+# 1. The solver is added to solvers/__init__.py
+#        __all__, and inference_tools_table
 # 2. The four class methods are implemented and they return correct data types
 #        required_in_problem, optional_in_problem
 #        required_in_options, optional_in_options