+++ tests/cofi_tools/test_scipy_lstsq.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from cofi.solvers import ScipyLstSqSolver
+from cofi.tools import ScipyLstSqSolver
 from cofi import BaseProblem, InversionOptions
 
 
@@ -17,6 +17,7 @@ def problem_setup():
     lamda = 1
     L = np.eye(2)
     return inv_problem, Cdinv, lamda, L
+
 
 def test_validate_jac():
     # 1
@@ -37,6 +38,7 @@ def test_validate_jac():
     with pytest.raises(ValueError, match=".*isn't set properly.*"):
         inv_solver = ScipyLstSqSolver(inv_problem, inv_options)
 
+
 def test_run(problem_setup):
     inv_problem, _, _, _ = problem_setup
     inv_options = InversionOptions()
@@ -45,6 +47,7 @@ def test_run(problem_setup):
     assert res["success"]
     assert res["model"][0] == pytest.approx(0, abs=0.1)
     assert res["model"][1] == pytest.approx(1, abs=0.1)
+
 
 def test_uncertainty_cov(problem_setup):
     inv_problem, Cdinv, _, _ = problem_setup
@@ -58,6 +61,7 @@ def test_uncertainty_cov(problem_setup):
     assert res["model"][0] == pytest.approx(0, abs=0.05)
     assert res["model"][1] == pytest.approx(1, abs=0.05)
 
+
 def test_uncertainty_cov_inv(problem_setup):
     inv_problem, Cdinv, _, _ = problem_setup
     # 2
@@ -70,11 +74,12 @@ def test_uncertainty_cov_inv(problem_setup):
     assert res["model"][0] == pytest.approx(0, abs=0.05)
     assert res["model"][1] == pytest.approx(1, abs=0.05)
     # 3
-    Cdinv[0,1] = 1
+    Cdinv[0, 1] = 1
     inv_problem.set_data_covariance_inv(Cdinv)
     solver = ScipyLstSqSolver(inv_problem, inv_options)
     _G = inv_problem.jacobian(1)
     assert np.array_equal(solver._a, _G.T @ Cdinv @ _G)
+
 
 def test_tikhonov(problem_setup):
     inv_problem, Cdinv, lamda, L = problem_setup