+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!--next-version-placeholder-->
 
+## v0.1.3.dev0 (not yet released)
+
+### CoFI Solvers -> CoFI Tools
+
+- [#110](https://github.com/inlab-geo/cofi/issues/110) `BaseSolver` -> `BaseInferenceTool`
+
 ## v0.1.2.dev25 (06/02/2023)
 
 ### CoFI Solvers